#32 [rename] Dashboard2 -> Dashboard로 변경, 그래프 Component 컨벤션에 맞춰 수정
