+++ package.json
@@ -21,6 +21,7 @@
   "dependencies": {
     "@popperjs/core": "2.11.8",
     "@stomp/stompjs": "^7.0.0",
+    "@toast-ui/chart": "^4.6.1",
     "axios": "^1.7.2",
     "bootstrap": "5.3.2",
     "chart.js": "4.4.0",
