+++ src/views/components/ColumnGraph.vue
