+++ src/views/components/GaugeGraph.vue
